+++ src/services/PeopleService.ts
@@ -5,22 +5,39 @@ import { IPeoplePickerEntity } from '../components/PeoplePicker/IPeoplePickerPro
 import { getSp } from '../config/pnp.config';
 import { generateImageUrl } from '../utils/generateImageUrl';
 
+/**
+ * Service class for interacting with SharePoint People Picker.
+ */
 class PeopleService {
   private sp: SPFI;
+
+  /**
+   * Initializes the PeopleService instance.
+   * @param {WebPartContext} context - The SharePoint WebPart context.
+   */
   constructor(context: WebPartContext) {
     this.sp = getSp(context);
   }
+
+  /**
+   * Restricts unvalidated email addresses from appearing in search results.
+   * @param {IBasePeoplePickerEntity} user - The user entity to validate.
+   * @returns {boolean} - Returns true if the user is validated, otherwise false.
+   */
   private restrictUnInvalidated(user: IBasePeoplePickerEntity): boolean {
     if (user.EntityData.PrincipalType) {
-      if (user.EntityData.PrincipalType === 'UNVALIDATED_EMAIL_ADDRESS') {
-        return false;
-      } else {
-        return true;
-      }
-    } else {
-      return true;
+      return user.EntityData.PrincipalType !== 'UNVALIDATED_EMAIL_ADDRESS';
     }
+    return true;
   }
+
+  /**
+   * Searches for users in SharePoint based on a query string.
+   * @param {WebPartContext} context - The SharePoint WebPart context.
+   * @param {string} query - The search query string.
+   * @param {number} [maximumSuggestions=25] - The maximum number of suggestions to return.
+   * @returns {Promise<IPeoplePickerEntity[]>} - A list of matched people entities.
+   */
   public async searchUser(
     context: WebPartContext,
     query: string,
@@ -35,6 +52,7 @@ class PeopleService {
       PrincipalType: PrincipalType.All,
       QueryString: query,
     });
+
     response.filter((value) => this.restrictUnInvalidated(value));
     return response.map((value) => {
       let image = '';
@@ -47,6 +65,12 @@ class PeopleService {
     });
   }
 
+  /**
+   * Resolves a single user based on a query string.
+   * @param {WebPartContext} context - The SharePoint WebPart context.
+   * @param {string} query - The user query string.
+   * @returns {Promise<IPeoplePickerEntity>} - The resolved user entity.
+   */
   public async resolveUser(
     context: WebPartContext,
     query: string
@@ -60,6 +84,7 @@ class PeopleService {
       PrincipalType: PrincipalType.All,
       QueryString: query,
     });
+
     let image = '';
     if (resolvedUser.Description && resolvedUser.Description !== '') {
       image = generateImageUrl(context, resolvedUser.Description);